nba_dfs/requirements.txt
# NBA DFS Comprehensive Model Requirements

# Core data science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Machine learning
xgboost>=2.0.0
lightgbm>=4.0.0
tensorflow>=2.13.0
torch>=2.0.0
catboost>=1.2.0

# Bayesian statistics
pymc>=5.0.0
arviz>=0.16.0

# Optimization (lineup builder)
pulp>=2.7.0
ortools>=9.7.0
cvxpy>=1.4.0

# NBA API
nba_api>=1.4.0

# Web scraping / data fetching
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
playwright>=1.40.0
httpx>=0.25.0
aiohttp>=3.9.0

# X / Twitter API v2 (beat writer news feed)
tweepy>=4.14.0

# Async
asyncio
aiofiles>=23.0.0

# Data storage
sqlalchemy>=2.0.0
sqlite3

# Fuzzy matching
fuzzywuzzy>=0.18.0
python-levenshtein>=0.21.0
rapidfuzz>=3.0.0

# Plotting / visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Scheduling / monitoring
schedule>=1.2.0
apscheduler>=3.10.0

# Environment / config
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
tqdm>=4.66.0
rich>=13.0.0
loguru>=0.7.0
tenacity>=8.2.0
cachetools>=5.3.0

# Statistical
statsmodels>=0.14.0
pingouin>=0.5.0
